refactor: move sendLog to xr.ts and enhance logging - Move sendLog function from utils.ts to xr.ts with encoded API key - Add sendLog wrapper method in rerouter.ts following same pattern as sendActivityLog - Enhance sendLog to include deviceId and licenseId in message - Add logging when screen is frozen for better monitoring - Remove redundant warning log in stopEmulator
